refactor: Добавлены запросы вещи и тесты 1.2
