+++ src/main/java/ru/practicum/shareit/booking/controller/BookingController.java
@@ -1,7 +1,7 @@
 package ru.practicum.shareit.booking.controller;
 
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.booking.BookingService;
+import ru.practicum.shareit.booking.service.BookingService;
 import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.BookingDtoFull;
