+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -2,6 +2,7 @@ package ru.practicum.shareit.user;
 
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.practicum.shareit.user.service.UserService;
 import ru.practicum.shareit.validation.Exist;
 import ru.practicum.shareit.validation.New;
 import ru.practicum.shareit.user.dto.UserDto;
