+++ src/main/resources/application.properties
@@ -8,10 +8,7 @@ logging.level.org.springframework.transaction=INFO
 logging.level.org.springframework.transaction.interceptor=TRACE
 logging.level.org.springframework.orm.jpa.JpaTransactionManager=DEBUG
 
-#---
-# TODO Append connection to DB
-#---
-spring.config.activate.on-profile=ci,test
+#spring.config.activate.on-profile=ci,test
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.url=jdbc:h2:mem:shareit
 spring.datasource.username=test
